requirements.txt
boto3>=1.26.0
requests>=2.28.0
ansible-core>=2.13.0
pytest>=7.3.1
pytest-cov>=4.1.0
pylint>=2.17.0
black>=23.3.0
isort>=5.12.0
bandit>=1.7.5
safety>=2.3.5 